Replace http-auth with passport equivalent.
Preferring the passport version per suggestion here:
https://github.com/18F/development-guide/tree/master/nodejs#authentication

We'll expand this to also check IP range.
