+++ package.json
@@ -55,10 +55,12 @@
     "cfenv": "^1.0.4",
     "express": "^4.14.1",
     "helmet": "^3.5.0",
-    "http-auth": "^3.1.1",
     "lru-cache": "^4.0.2",
     "morgan": "^1.8.1",
     "newrelic": "^1.38.1",
+    "passport": "^0.3.2",
+    "passport-custom": "^1.0.5",
+    "passport-http": "^0.3.0",
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
     "react-resolver": "^3.1.0",
